Implement action creators to fetch data from the api
